+++ src/shared/constants/auction-status.ts
@@ -1,0 +1,5 @@
+export enum AuctionStatus {
+  Pending = 'pending',
+  Approved = 'approved',
+  Completed = 'completed',
+}
