+++ src/shared/constants/product-status.ts
@@ -1,0 +1,5 @@
+export enum ProductStatus {
+  Live = 'live',
+  Sold = 'sold',
+  Delivered = 'delivered',
+}
