+++ src/shared/constants/user-role.enum.ts
@@ -1,0 +1,5 @@
+export enum UserRole {
+  Admin = 'admin',
+  Seller = 'seller',
+  Buyer = 'buyer',
+}
