+++ src/user/role/role.decorator.ts
@@ -1,6 +1,6 @@
 import { SetMetadata } from '@nestjs/common';
-import { UserRole } from '../schemas/user.schema';
 import { createParamDecorator, ExecutionContext } from '@nestjs/common';
+import { UserRole } from 'src/shared/constants/user-role.enum';
 
 export const Roles = (...roles: UserRole[]) => SetMetadata('roles', roles);
